Thu, Aug 10, 2023, 6:45 PM +09:00

+++ other/average.c
@@ -3,7 +3,7 @@
 
 const int N = 3; //the length of your array; how many numbers do you want the average of?
 
-int get_average(int length, int array[]);
+float get_average(int length, int array[]);
 
 int main (void)
 {
@@ -13,10 +13,10 @@ int main (void)
     {
         numbers[i] = get_int("Write a number: ");
     }
-    printf("The average of the %i numbers is: %d\n", N, get_average(N, numbers));
+    printf("The average of the %i numbers is: %f\n", N, get_average(N, numbers));
 }
 
-int get_average(int length, int array[])
+float get_average(int length, int array[])
 {
     int sum = 0;
     for (int i = 0; i < length; i++)
